adjustments, please read desc:
So I adjusted the code of both cat and dog and decided to make them to share the same adoption screen. If you want, you can change the adoption color because it looks more like focused on cat. I also made it so the dog and cat detail screen are looking the same, but with different color palettes to match our original code. I added the letter thingy Nancy had for dog API into cat to match, and back buttons for all 4 screens.